[Task - See 'WEBSITE-SCRAPING-Truthfinder.com' mindmap](https://drive.google.com/file/d/16LXDHxf6q1Repj1E2KlCpvZRVnLClDjK/view?usp=drive_link) (['mount path'](google-drive://mayou.nkensa@gmail.com/GVfsSharedWithMe/0B39Sj9mqtvSyM2NlZjZjZWYtNGVjZC00NTU0LTkzMWUtNGIzZjVmNTgyMTc4/0B39Sj9mqtvSycEVjTV9pZllfUzQ/0B39Sj9mqtvSyTzdtakJ1RW41am8/1cVkUyxtFhjCWj-D5u--Q8AwOqpSfC51J/1Qm18Vv2T-N092weQNiCMIISKS7lga7hr/1I17iJ_UagHNNPgQTM8LeZbaQUQCU7kI_/1y4aqHqRC3lLIkkz6FYHNYOuv8TJRvsPy/1JDBULfFjCeXVbU98R0qMj_kbA4ZQCFZa/1Qnaumo16wPlw82PVMOApqMURsYqJFeKQ/1axsRJhxUnUr17IQUkIzy70XD_cN4QFE7) - Webscraper demo implemented | Refactoring in place for config controlled process | TESTED, ADJUSTED, OK... | First full webscrape and webscrapes batch implemented and tested, to be reviewed... | Corrections + Adding firstName+lastName inputs sequenced + assertion for process interruption (SAFEKEEPING, not validated) | Adjustments, little cleanup... | Taskhandling implementation :: (1. cache, 2. create task handler, 3. create webscrape method) done → Next: 4. create api methods, 5. test, validate...| 12th Done → Deploying...

+++ django-unicorn/webscraping/views.py
@@ -108,7 +108,7 @@ def webscrape_long_running_method( webscrape: Webscrape, task_progress ):
             progress_message=f'Sequence {i} has been processed | Details: {webscrape}' )
 
         if i > 0:
-            webscrape.task_progress = ( i / sequences_len) * 100
+            webscrape.task_progress = int(( i / sequences_len) * 100)
             webscrape.save()
 
     driver.close()
